+++ src/imageclasses/ImgRaster.h
@@ -78,8 +78,8 @@ public:
   ///constructor opening an image in memory using an external data pointer (not tested yet)
   //ImgRaster(void* dataPointer, unsigned int ncol, unsigned int nrow, int nband, const GDALDataType& dataType);
   //from Reader
-  ImgRaster(const std::string& filename, unsigned long int memory=0) : ImgRaster() {open(filename, memory);};
-  // ImgRaster(const std::string& filename, const GDALAccess& readMode=GA_ReadOnly, unsigned long int memory=0) : m_writeMode(false) {open(filename, readMode, memory);};
+  ImgRaster(const std::string& filename, unsigned int memory=0) : ImgRaster() {open(filename, memory);};
+  // ImgRaster(const std::string& filename, const GDALAccess& readMode=GA_ReadOnly, unsigned int memory=0) : m_writeMode(false) {open(filename, readMode, memory);};
   //from Writer
   ///constructor opening an image for writing, copying image attributes from a source image. Caching is supported when memory>0
   ImgRaster(const std::string& filename, const ImgRaster& imgSrc, unsigned int memory=0, const std::vector<std::string>& options=std::vector<std::string>()) : ImgRaster() {open(filename, imgSrc, memory, options);};
@@ -96,7 +96,7 @@ public:
   virtual ~ImgRaster(void){freeMem();};
 
   ///Initialize the memory for read/write image in cache
-  void initMem(unsigned long int memory);
+  void initMem(unsigned int memory);
   ///assignment operator
   ImgRaster& operator=(ImgRaster& imgSrc);
   ///get write mode
@@ -268,8 +268,8 @@ public:
 
   //From Reader
   ///Open an image. Set memory (in MB) to cache a number of rows in memory
-  void open(const std::string& filename, unsigned long int memory=0);
-  // void open(const std::string& filename, const GDALAccess& readMode=GA_ReadOnly, unsigned long int memory=0);
+  void open(const std::string& filename, unsigned int memory=0);
+  // void open(const std::string& filename, const GDALAccess& readMode=GA_ReadOnly, unsigned int memory=0);
   ///Read all pixels from image in memory for specific band
   CPLErr readData(unsigned int band);
   ///Read all pixels from image in memory
@@ -337,9 +337,9 @@ public:
   ///Write pixel cell values for a range of columns and rows for a specific band (all indices start counting from 0). The buffer is a two dimensional vector (stl vector of stl vector) representing [row][col].
   template<typename T> CPLErr writeDataBlock(Vector2d<T>& buffer2d, unsigned int minCol, unsigned int maxCol, int minRow, int maxRow, unsigned int band=0);
   ///Prepare image writer to write to file
-  void setFile(const std::string& filename, const std::string& imageType, unsigned long int memory=0, const std::vector<std::string>& options=std::vector<std::string>());
+  void setFile(const std::string& filename, const std::string& imageType, unsigned int memory=0, const std::vector<std::string>& options=std::vector<std::string>());
   ///Prepare image writer to write to file
-  // void setFile(const std::string& filename, const ImgRaster& imgSrc, unsigned long int memory=0, const std::vector<std::string>& options=std::vector<std::string>());
+  // void setFile(const std::string& filename, const ImgRaster& imgSrc, unsigned int memory=0, const std::vector<std::string>& options=std::vector<std::string>());
   ///Set the color table using an (ASCII) file with 5 columns (value R G B alpha)
   void setColorTable(const std::string& filename, unsigned int band=0);
   ///Set the color table using the GDAL class GDALColorTable
